split global var manager
